+++ history.md
@@ -1,5 +1,8 @@
 ### History
 
+dev
+- Centralize version and date in project init.
+
 v2.1.1
 - Disable automatic symbology application when opening the plugin.
 - Update tooltips of GUI elements.