+++ history.md
@@ -3,6 +3,7 @@
 dev
 - Improve grd file handling.
 - Update reference paper details.
+- Centralize version and date in project init.
 
 v2.1.1
 - Disable automatic symbology application when opening the plugin.